+++ src/components/spectate/components/visits-table/visits-table.component.js
@@ -1,11 +1,11 @@
 module.exports = {
     onCreate(input) {
         this.state = {
-            visits: input.visits
+            visits: input.visits.reverse()
         }
     },
     setVisits(visits) {
-        this.state.visits = visits;
+        this.state.visits = visits.reverse();
         this.setStateDirty('visits');
     }
 }